resolve dependency issues for imports
