+++ azurelinuxagent/common/event.py
@@ -28,9 +28,8 @@ import azurelinuxagent.common.conf as conf
 import azurelinuxagent.common.logger as logger
 from azurelinuxagent.common.exception import EventError
 from azurelinuxagent.common.future import ustr
-from azurelinuxagent.common.protocol.restapi import TelemetryEventParam, \
-    TelemetryEvent, \
-    get_properties
+from azurelinuxagent.common.datacontract import get_properties
+from azurelinuxagent.common.telemetryevent import TelemetryEventParam, TelemetryEvent
 from azurelinuxagent.common.utils import fileutil, textutil
 from azurelinuxagent.common.version import CURRENT_VERSION
