+++ tests/protocol/test_imds.py
@@ -1,12 +1,28 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
+# Microsoft Azure Linux Agent
+#
+# Copyright 2018 Microsoft Corporation
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Requires Python 2.6+ and Openssl 1.0+
+
 import json
 
 import azurelinuxagent.common.protocol.imds as imds
 
+from azurelinuxagent.common.datacontract import set_properties
 from azurelinuxagent.common.exception import HttpError
 from azurelinuxagent.common.future import ustr
-from azurelinuxagent.common.protocol.restapi import set_properties
 from azurelinuxagent.common.utils import restutil
 from tests.ga.test_update import ResponseMock
 from tests.tools import *
